ml-service/requirements.txt
# Deep Learning Frameworks
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
keras==2.15.0

# YOLO
ultralytics==8.0.227
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# ML & Data Processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4

# Time Series
statsmodels==0.14.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Image Processing
Pillow==10.1.0
imageio==2.33.0

# Model Serialization
joblib==1.3.2
onnx==1.15.0
onnxruntime==1.16.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# GPU Support (optional)
# torch-cuda==12.1  # Uncomment if using CUDA
